requirements.txt
locust==1.4.3
beautifulsoup4==4.9.3
python-decouple==3.4